+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Changed
 * Include all necessary files for running the tests in source distribution
 * Exclude tests from built/installed version of the package
 * File field passed to ``run`` function can be url address
+* Connect to a local server as public user by default
 
 Fixed
 -----